Test for /var/tmp's existence

+++ root/etc/orchestration/start.sh
@@ -1,7 +1,12 @@
 #!/bin/bash
 
 # Cleanup temp files
-find /var/tmp/ -exec rm -rf {} \;
+if [ -d "/var/tmp" ]; then
+    echo "Deleting contents of /var/tmp..."
+    find /var/tmp/ -exec rm -rf {} \;
+else
+    echo "/var/tmp doesn't exist; nothing to cleanup."
+fi
 
 # Begin OpenVPN initialization
 . /etc/openvpn/init.sh
